pyproject.toml
[build-system]
requires = ["hatchling>=1.25"]
build-backend = "hatchling.build"

[project]
name = "url-shortener-api"
version = "0.1.0"
description = "Scalable FastAPI URL shortener with MongoDB, Redis caching, Celery analytics, and JWT authentication."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110,<0.120",
    "uvicorn[standard]>=0.23,<0.31",
    "motor>=3.3,<4.0",
    "pydantic>=2.6,<3.0",
    "pydantic-settings>=2.3,<3.0",
    "passlib[bcrypt]>=1.7,<2.0",
    "python-jose[cryptography]>=3.3,<4.0",
    "bcrypt>=4.0,<5.0",
    "redis>=5.0,<6.0",
    "celery[redis]>=5.3,<6.0",
    "structlog>=24.1,<25.0",
    "python-slugify>=8.0,<9.0",
    "email-validator>=2.1,<3.0",
    "shortuuid>=1.0,<2.0",
    "aiofiles>=23.2,<24.0",
    "httpx>=0.27,<0.29",
    "uvloop>=0.19; sys_platform != 'win32'",
    "async-timeout>=4.0,<5.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4,<8.0",
    "pytest-asyncio>=0.23,<0.24",
    "httpx>=0.27,<0.29",
    "coverage>=7.4,<8.0",
    "ruff>=0.5,<0.6"
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "UP", "ASYNC"]
